Luca main

+++ sw/applications/0_DOOM/doom/r_draw.c
@@ -360,7 +360,7 @@ const int     fuzzoffset[FUZZTABLE] =
     FUZZOFF,FUZZOFF,-FUZZOFF,FUZZOFF,FUZZOFF,-FUZZOFF,FUZZOFF 
 }; 
 
-int     fuzzpos = 0; 
+uint8_t     fuzzpos = 0; 
 
 
 //
